work in progress

+++ cellular_controller/helpers/cellular_helpers.c
@@ -120,7 +120,7 @@ int8_t socket_connect(struct data *data, struct sockaddr *addr,
 	return ret;
 }
 
-int8_t socket_listen(struct data *data, uint16_t port)
+int socket_listen(struct data *data, uint16_t port)
 {
 	int ret;
 	struct sockaddr_in bind_addr;
@@ -146,7 +146,7 @@ int8_t socket_listen(struct data *data, uint16_t port)
 //	}
 	k_sleep(K_MSEC(50));
 	ret = listen(data->tcp.sock, 1); //2nd parameter is backlog size, not
-	// important as we are not interested in the incoming data.
+	// important as we are not interested in the incoming data anyways.
 	if (ret < 0) {
 		LOG_ERR("Cannot start TCP listening socket (%s): %d",
 			data->proto,
